Update archetype to reflect fixes to the quickstart of the same code base.
