+++ spring-mvc-archetype/pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>org.jboss.spring.archetypes</groupId>
     <artifactId>jboss-spring-mvc-archetype</artifactId>
-    <version>1.0.1.SNAPSHOT</version>
+    <version>1.0.1-SNAPSHOT</version>
     <packaging>maven-archetype</packaging>
 
     <name>JBoss AS7 - Spring MVC Application</name>
