+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/java/controller/MemberController.java
@@ -1,14 +1,15 @@
 #set( $symbol_pound = '#' )
 #set( $symbol_dollar = '$' )
 #set( $symbol_escape = '\' )
-package ${package}.mvc;
+package ${package}.controller;
 
- import javax.validation.Valid;
+import javax.validation.Valid;
 
-import ${package}.domain.Member;
-import ${package}.repo.MemberDao;
+import ${package}.data.MemberDao;
+import ${package}.model.Member;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.UnexpectedRollbackException;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -16,26 +17,30 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping(value="/")
+@RequestMapping(value = "/")
 public class MemberController {
     @Autowired
     private MemberDao memberDao;
 
-    @RequestMapping(method=RequestMethod.GET)
-    public String displaySortedMembers(Model model)
-    {
+    @RequestMapping(method = RequestMethod.GET)
+    public String displaySortedMembers(Model model) {
         model.addAttribute("newMember", new Member());
         model.addAttribute("members", memberDao.findAllOrderedByName());
         return "index";
     }
 
-    @RequestMapping(method=RequestMethod.POST)
+    @RequestMapping(method = RequestMethod.POST)
     public String registerNewMember(@Valid @ModelAttribute("newMember") Member newMember, BindingResult result, Model model) {
         if (!result.hasErrors()) {
-            memberDao.register(newMember);
-            return "redirect:/";
-        }
-        else {
+            try {
+                memberDao.register(newMember);
+                return "redirect:/";
+            } catch (UnexpectedRollbackException e) {
+                model.addAttribute("members", memberDao.findAllOrderedByName());
+                model.addAttribute("error", e.getCause().getCause());
+                return "index";
+            }
+        } else {
             model.addAttribute("members", memberDao.findAllOrderedByName());
             return "index";
         }
