+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/java/controller/MemberRestController.java
@@ -1,12 +1,12 @@
 #set( $symbol_pound = '#' )
 #set( $symbol_dollar = '$' )
 #set( $symbol_escape = '\' )
-package ${package}.mvc;
+package ${package}.controller;
 
 import java.util.List;
 
-import ${package}.domain.Member;
-import ${package}.repo.MemberDao;
+import ${package}.data.MemberDao;
+import ${package}.model.Member;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,16 +17,17 @@ import org.springframework.web.bind.annotation.ResponseBody;
 @Controller
 @RequestMapping("/rest/members")
 public class MemberRestController {
-
     @Autowired
     private MemberDao memberDao;
 
-    @RequestMapping(method=RequestMethod.GET, produces="application/json")
-    public @ResponseBody List<Member> listAllMembers() {
+    @RequestMapping(method = RequestMethod.GET, produces = "application/json")
+    public
+    @ResponseBody
+    List<Member> listAllMembers() {
         return memberDao.findAllOrderedByName();
     }
 
-    @RequestMapping(value="/{id}", method=RequestMethod.GET, produces="application/json")
+    @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = "application/json")
     public @ResponseBody Member lookupMemberById(@PathVariable("id") Long id) {
         return memberDao.findById(id);
     }
