+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/java/data/MemberDao.java
@@ -1,11 +1,11 @@
 #set( $symbol_pound = '#' )
 #set( $symbol_dollar = '$' )
 #set( $symbol_escape = '\' )
-package ${package}.repo;
+package ${package}.data;
 
 import java.util.List;
 
-import ${package}.domain.Member;
+import ${package}.model.Member;
 
 public interface MemberDao {
     public Member findById(Long id);
