+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/resources/META-INF/persistence.xml
@@ -2,21 +2,22 @@
 #set( $symbol_dollar = '$' )
 #set( $symbol_escape = '\' )
 <persistence version="2.0"
-   xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="
+    xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="
         http://java.sun.com/xml/ns/persistence
         http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd">
-   <persistence-unit name="primary">
-      <!-- If you are running in a production environment, add a managed 
-         data source, this example data source is just for development and testing! -->
-      <!-- The datasource is deployed as WEB-INF/spring-quickstart-ds.xml, you
-         can find it in the source at src/main/webapp/WEB-INF/spring-quickstart-ds.xml -->
-      <jta-data-source>java:jboss/datasources/SpringQuickstartDS</jta-data-source>
-      <properties>
-		 <property name="jboss.entity.manager.factory.jndi.name" value="java:jboss/spring-quickstart/persistence" />
-         <!-- Properties for Hibernate -->
-         <property name="hibernate.hbm2ddl.auto" value="create-drop" />
-         <property name="hibernate.show_sql" value="false" />
-      </properties>
-   </persistence-unit>
+    <persistence-unit name="primary">
+        <!-- If you are running in a production environment, add a managed 
+           data source, this example data source is just for development and testing! -->
+        <!-- The datasource is deployed as WEB-INF/spring-quickstart-ds.xml, you
+           can find it in the source at src/main/webapp/WEB-INF/spring-quickstart-ds.xml -->
+        <jta-data-source>java:jboss/datasources/SpringQuickstartDS</jta-data-source>
+        <properties>
+            <property name="jboss.entity.manager.factory.jndi.name" 
+                      value="java:jboss/spring-quickstart/persistence" />
+            <!-- Properties for Hibernate -->
+            <property name="hibernate.hbm2ddl.auto" value="create-drop" />
+            <property name="hibernate.show_sql" value="false" />
+        </properties>
+    </persistence-unit>
 </persistence>
