+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/resources/META-INF/spring/infrastructure.xml
@@ -4,27 +4,23 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <beans xmlns="http://www.springframework.org/schema/beans"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       	 xmlns:jee="http://www.springframework.org/schema/jee"
-       	 xmlns:tx="http://www.springframework.org/schema/tx"
-         xsi:schemaLocation="http://www.springframework.org/schema/beans
-           http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-           http://www.springframework.org/schema/tx
-		   http://www.springframework.org/schema/tx/spring-tx.xsd
-		   http://www.springframework.org/schema/jee 
-		   http://www.springframework.org/schema/jee/spring-jee.xsd">
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:jee="http://www.springframework.org/schema/jee"
+       xmlns:tx="http://www.springframework.org/schema/tx"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
+                           http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
+                           http://www.springframework.org/schema/jee http://www.springframework.org/schema/jee/spring-jee.xsd">
 
-	<!-- JDNI name for EntityManagerFactory is defined in src/main/resources/META-INF/persistence.xml -->
+    <!-- JDNI name for EntityManagerFactory is defined in src/main/resources/META-INF/persistence.xml -->
     <jee:jndi-lookup jndi-name="java:jboss/spring-quickstart/persistence" id="entityManagerFactory"
                      expected-type="javax.persistence.EntityManagerFactory" />
 
-	<bean id="entityManager" class="org.springframework.orm.jpa.support.SharedEntityManagerBean">
-   		<property name="entityManagerFactory" ref="entityManagerFactory"/>
-	</bean>
-                     
-	<jee:jndi-lookup jndi-name="java:jboss/datasources/SpringQuickstartDS" id="dataSource" expected-type="javax.sql.DataSource"/>
+    <bean id="entityManager" class="org.springframework.orm.jpa.support.SharedEntityManagerBean">
+        <property name="entityManagerFactory" ref="entityManagerFactory" />
+    </bean>
+
+    <jee:jndi-lookup jndi-name="java:jboss/datasources/SpringQuickstartDS" id="dataSource" expected-type="javax.sql.DataSource" />
 
     <tx:jta-transaction-manager />
 
-	
 </beans>
