+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/webapp/WEB-INF/jboss-deployment-structure.xml
@@ -1,10 +1,10 @@
 <jboss-deployment-structure xmlns="urn:jboss:deployment-structure:1.0">
-   <deployment>
-   	   <dependencies>
-   	   		<module name="com.h2database.h2"/>
-   	   		<module name="org.codehaus.jackson.jackson-core-asl"/>
-   	   		<module name="org.codehaus.jackson.jackson-mapper-asl"/>
+    <deployment>
+        <dependencies>
+            <module name="com.h2database.h2"/>
+            <module name="org.codehaus.jackson.jackson-core-asl"/>
+            <module name="org.codehaus.jackson.jackson-mapper-asl"/>
             <module name="org.slf4j"/>
-   	   </dependencies>
-   </deployment>
+        </dependencies>
+    </deployment>
 </jboss-deployment-structure>
