+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/webapp/WEB-INF/spring-quickstart-ds.xml
@@ -3,19 +3,19 @@
    or testing only. It uses H2, an in memory database that ships with JBoss 
    AS. -->
 <datasources xmlns="http://www.jboss.org/ironjacamar/schema"
-   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-   xsi:schemaLocation="http://www.jboss.org/ironjacamar/schema http://docs.jboss.org/ironjacamar/schema/datasources_1_0.xsd">
-   <!-- The datasource is bound into JNDI at this location. We reference 
-      this in META-INF/persistence.xml -->
-   <datasource jndi-name="java:jboss/datasources/SpringQuickstartDS"
-      pool-name="kitchensink-quickstart" enabled="true"
-      use-java-context="true">
-      <connection-url>jdbc:h2:mem:kitchensink-quickstart;DB_CLOSE_DELAY=-1</connection-url>
-      <driver>h2</driver>
-      <security>
-         <user-name>sa</user-name>
-         <password>sa</password>
-      </security>
-   </datasource>
+             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+             xsi:schemaLocation="http://www.jboss.org/ironjacamar/schema http://docs.jboss.org/ironjacamar/schema/datasources_1_0.xsd">
+    <!-- The datasource is bound into JNDI at this location. We reference
+       this in META-INF/persistence.xml -->
+    <datasource jndi-name="java:jboss/datasources/SpringQuickstartDS"
+                pool-name="kitchensink-quickstart" enabled="true"
+                use-java-context="true">
+        <connection-url>jdbc:h2:mem:kitchensink-quickstart;DB_CLOSE_DELAY=-1</connection-url>
+        <driver>h2</driver>
+        <security>
+            <user-name>sa</user-name>
+            <password>sa</password>
+        </security>
+    </datasource>
 </datasources>
