+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/main/webapp/WEB-INF/web.xml
@@ -5,33 +5,34 @@
          xmlns="http://java.sun.com/xml/ns/javaee"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://java.sun.com/xml/ns/javaee 
-         http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" 
+         http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
          metadata-complete="true">
 
- <display-name>Java EE 6 Starter Application</display-name>
+    <display-name>Java EE 6 Starter Application</display-name>
 
-  <context-param>
-	<param-name>contextConfigLocation</param-name>
-	<param-value>classpath:/META-INF/spring/applicationContext.xml,
-				classpath:/META-INF/spring/infrastructure.xml</param-value>
-  </context-param>
+    <context-param>
+        <param-name>contextConfigLocation</param-name>
+        <param-value>classpath:/META-INF/spring/applicationContext.xml,
+            classpath:/META-INF/spring/infrastructure.xml
+        </param-value>
+    </context-param>
 
-  <listener>
-	<listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
-  </listener>
+    <listener>
+        <listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
+    </listener>
 
-  <servlet>
-	<servlet-name>jboss-as-kitchensink</servlet-name>
-	<servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
-	<init-param>
-	  <param-name>contextConfigLocation</param-name>
-	  <param-value>/WEB-INF/jboss-as-spring-mvc-context.xml</param-value>
-	</init-param>
-	<load-on-startup>1</load-on-startup>
-  </servlet>
+    <servlet>
+        <servlet-name>jboss-as-kitchensink</servlet-name>
+        <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
+        <init-param>
+            <param-name>contextConfigLocation</param-name>
+            <param-value>/WEB-INF/jboss-as-spring-mvc-context.xml</param-value>
+        </init-param>
+        <load-on-startup>1</load-on-startup>
+    </servlet>
 
-  <servlet-mapping>
-	<servlet-name>jboss-as-kitchensink</servlet-name>
-	<url-pattern>/</url-pattern>
-  </servlet-mapping>
+    <servlet-mapping>
+        <servlet-name>jboss-as-kitchensink</servlet-name>
+        <url-pattern>/</url-pattern>
+    </servlet-mapping>
 </web-app>
