+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/test/resources/META-INF/test-persistence.xml
@@ -4,12 +4,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <persistence xmlns="http://java.sun.com/xml/ns/persistence" version="1.0">
-	<persistence-unit name="testPU" transaction-type="RESOURCE_LOCAL">
+    <persistence-unit name="testPU" transaction-type="RESOURCE_LOCAL">
         <provider>org.hibernate.ejb.HibernatePersistence</provider>
-		<class>${package}.domain.Member</class>
+        <class>${package}.model.Member</class>
         <properties>
-	        <property name="hibernate.hbm2ddl.auto" value="create-drop" />
-        	<property name="hibernate.show_sql" value="true" />
+            <property name="hibernate.hbm2ddl.auto" value="create-drop"/>
+            <property name="hibernate.show_sql" value="true"/>
         </properties>
-	</persistence-unit>
+    </persistence-unit>
 </persistence>
