+++ spring-mvc-archetype/src/main/resources/archetype-resources/src/test/resources/import.sql
@@ -1,2 +1,2 @@
 -- You can use this file to load seed data into the database using SQL statements
-insert into Member (id, name, email, phone_number) values (0, 'John Smith', 'john.smith@mailinator.com', '2125551212') 
+insert into Member (id, name, email, phone_number) values (0, 'John Smith', 'john.smith@mailinator.com', '2125551212')
